feat(order): change SequenceNumber from int to long
Also removes explicit cast on Redis INCR result and (int?) cast for
MaxAsync. Migration regenerated with bigint for SequenceNumber column.

+++ backend/Dukaan/Dukaan.Application/Interfaces/IOrderNumberService.cs
@@ -2,5 +2,5 @@ namespace Dukaan.Application.Interfaces;
 
 public interface IOrderNumberService
 {
-    Task<(int SequenceNumber, string OrderNumber)> GetNextOrderNumberAsync(CancellationToken cancellationToken = default);
+    Task<(long SequenceNumber, string OrderNumber)> GetNextOrderNumberAsync(CancellationToken cancellationToken = default);
 }
